feat(rm): build qagame as 64-bit DLL (module-side intptr_t ABI)
Completes the VM ABI port on the module side and adds the qagame_mp_x86_64.dll
target (game/ + botai/, -DGAMEDLL, exported via game.def). Builds clean 64-bit.

- g_syscalls.c: syscall ptr + dllEntry widened to intptr_t (matches engine).
- g_main.c: vmMain returns intptr_t; ClientConnect denial-string return cast
  to intptr_t. Known latent (bot-only, compiles w/ warnings): BOT_VISIBLEFROMPOS
  / BOT_CHECKATTACKATPOS pass vec3 pointers through vmMain int args (truncates;
  only hit with bots active) - to fix when wiring bot support.

Build milestone only: all of etrm/etrmded/qagame compile+link 64-bit. Runtime
still crashes at init (ntdll heap access-violation 0xC0000005, both binaries) -
a 64-bit memory bug in early init, reachable now that 2.60b paks are present.
Next: AddressSanitizer to pinpoint it.

+++ src/game/g_main.c
@@ -479,7 +479,7 @@ This must be the very first function compiled into the .q3vm file
 #if __GNUC__ >= 4
 #pragma GCC visibility push(default)
 #endif
-int vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6 ) {
+intptr_t vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6 ) {
 #if __GNUC__ >= 4
 #pragma GCC visibility pop
 #endif
@@ -491,7 +491,7 @@ int vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int a
 		G_ShutdownGame( arg0 );
 		return 0;
 	case GAME_CLIENT_CONNECT:
-		return (int)ClientConnect( arg0, arg1, arg2 );
+		return (intptr_t)ClientConnect( arg0, arg1, arg2 );   // returns denial-reason string ptr
 	case GAME_CLIENT_THINK:
 		ClientThink( arg0 );
 		return 0;
